feat: complete architecture overhaul, Redis event queue, DB indexes, and Android on-device datapackage caching
