+++ AGENTS.md
@@ -1,0 +1,6 @@
+# Agent Orientation Guide — Archipelago Alerts
+
+This repository contains:
+- **[LLM.md](LLM.md):** Complete codebase reference, data models, gotchas, and guidelines for AI agents.
+- **[architecture.md](architecture.md):** Detailed system architecture, Mermaid diagrams, Redis Pub/Sub events, and containerization structure.
+- **[CLAUDE.md](CLAUDE.md):** Quick command reference for Claude Code CLI and coding assistants.
