+++ alembic/versions/a971160f9c05_add_filler_trap_prefs.py
@@ -22,10 +22,10 @@ def upgrade() -> None:
     """Upgrade schema."""
     # ### commands auto generated by Alembic - please adjust! ###
     with op.batch_alter_table('users', schema=None) as batch_op:
-        batch_op.add_column(sa.Column('notify_filler_default', sa.Boolean(), nullable=False, server_default=sa.text('0')))
-        batch_op.add_column(sa.Column('notify_trap_default', sa.Boolean(), nullable=False, server_default=sa.text('0')))
-        batch_op.add_column(sa.Column('ui_show_filler_default', sa.Boolean(), nullable=False, server_default=sa.text('0')))
-        batch_op.add_column(sa.Column('ui_show_trap_default', sa.Boolean(), nullable=False, server_default=sa.text('0')))
+        batch_op.add_column(sa.Column('notify_filler_default', sa.Boolean(), nullable=False, server_default=sa.text('false')))
+        batch_op.add_column(sa.Column('notify_trap_default', sa.Boolean(), nullable=False, server_default=sa.text('false')))
+        batch_op.add_column(sa.Column('ui_show_filler_default', sa.Boolean(), nullable=False, server_default=sa.text('false')))
+        batch_op.add_column(sa.Column('ui_show_trap_default', sa.Boolean(), nullable=False, server_default=sa.text('false')))
 
     with op.batch_alter_table('user_tracked_slots', schema=None) as batch_op:
         batch_op.add_column(sa.Column('notify_filler', sa.Boolean(), nullable=True))
