+++ android/app/src/main/java/com/jones/aptracker/database/AppDatabase.kt
@@ -11,12 +11,13 @@ import com.jones.aptracker.network.HistoryItemEntity
 import com.jones.aptracker.network.RoomDao
 import com.jones.aptracker.network.RoomEntity
 
-@Database(entities = [RoomEntity::class, HistoryItemEntity::class, HintEntity::class], version = 18)
+@Database(entities = [RoomEntity::class, HistoryItemEntity::class, HintEntity::class, CachedDatapackageEntity::class], version = 20)
 abstract class AppDatabase : RoomDatabase() {
 
     abstract fun roomDao(): RoomDao
     abstract fun historyDao(): HistoryDao
     abstract fun hintDao(): HintDao
+    abstract fun datapackageDao(): DatapackageDao
 
     companion object {
         @Volatile
@@ -32,7 +33,7 @@ abstract class AppDatabase : RoomDatabase() {
                     .addMigrations(
                         MIGRATION_5_6, MIGRATION_6_7, MIGRATION_7_8, MIGRATION_8_9,
                         MIGRATION_9_10, MIGRATION_10_11, MIGRATION_11_12, MIGRATION_12_13, MIGRATION_13_14, MIGRATION_14_15,
-                        MIGRATION_15_16, MIGRATION_16_17, MIGRATION_17_18
+                        MIGRATION_15_16, MIGRATION_16_17, MIGRATION_17_18, MIGRATION_18_19, MIGRATION_19_20
                     )
                     .fallbackToDestructiveMigration()
                     .build()
