+++ backend/requirements.txt
@@ -51,9 +51,12 @@ requests==2.32.5
 rsa==4.9.1
 sniffio==1.3.1
 SQLAlchemy==2.0.43
+psycopg2-binary==2.9.12
+redis==8.0.1
 typing_extensions==4.15.0
 urllib3==2.5.0
 waitress==3.0.2
 websockets==15.0.1
 Werkzeug==3.1.3
 yarl==1.22.0
+
