+++ docker-compose.dev.yml
@@ -1,0 +1,23 @@
+version: '3.8'
+
+services:
+  postgres:
+    image: postgres:15-alpine
+    container_name: ap_tracker_db_dev
+    environment:
+      POSTGRES_USER: ap_user
+      POSTGRES_PASSWORD: ap_password
+      POSTGRES_DB: ap_tracker_dev
+    ports:
+      - "5432:5432"
+    volumes:
+      - pgdata_dev:/var/lib/postgresql/data
+
+  redis:
+    image: redis:7-alpine
+    container_name: ap_tracker_redis_dev
+    ports:
+      - "6379:6379"
+
+volumes:
+  pgdata_dev:
